+++ tests/test_twisted.py
@@ -11,6 +11,8 @@ from twisted.internet import ssl
 
 import pem
 
+from pem.twisted import certificateOptionsFromFiles
+
 from .data import KEY_PEM, KEY_PEM2, CERT_PEMS
 
 
@@ -30,7 +32,7 @@ class TestCertificateOptionsFromFiles(object):
         keyFile.write(KEY_PEM)
         certFile = tmpdir.join('cert.pem')
         certFile.write(CERT_PEMS[0])
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(keyFile), str(certFile),
         )
         assert [] == ctxFactory.extraCertChain
@@ -42,7 +44,7 @@ class TestCertificateOptionsFromFiles(object):
         certFile.write(CERT_PEMS[0])
         chainFile = tmpdir.join('chain.pem')
         chainFile.write(''.join(CERT_PEMS[1:]))
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(keyFile), str(certFile), str(chainFile)
         )
         assert 2 == len(ctxFactory.extraCertChain)
@@ -52,7 +54,7 @@ class TestCertificateOptionsFromFiles(object):
         keyFile.write(KEY_PEM)
         certFile = tmpdir.join('cert_and_chain.pem')
         certFile.write(''.join(CERT_PEMS))
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(keyFile), str(certFile)
         )
         assert 2 == len(ctxFactory.extraCertChain)
@@ -67,24 +69,24 @@ class TestCertificateOptionsFromFiles(object):
         keyFile.write(KEY_PEM)
         certFile = tmpdir.join('cert_and_chain.pem')
         certFile.write(''.join(reversed(CERT_PEMS)))
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(keyFile), str(certFile)
         )
         assert 2 == len(ctxFactory.extraCertChain)
 
     def test_worksWithEverythingInOneFile(self, allFile):
-        ctxFactory = pem.certificateOptionsFromFiles(str(allFile))
+        ctxFactory = certificateOptionsFromFiles(str(allFile))
         assert 2 == len(ctxFactory.extraCertChain)
 
     def test_passesCertsInCorrectFormat(self, allFile):
-        ctxFactory = pem.certificateOptionsFromFiles(str(allFile))
+        ctxFactory = certificateOptionsFromFiles(str(allFile))
         assert isinstance(ctxFactory.privateKey, crypto.PKey)
         assert isinstance(ctxFactory.certificate, crypto.X509)
         assert all(isinstance(cert, crypto.X509)
                    for cert in ctxFactory.extraCertChain)
 
     def test_forwardsKWargs(self, allFile):
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(allFile),
             method=SSL.TLSv1_METHOD,
         )
@@ -94,7 +96,7 @@ class TestCertificateOptionsFromFiles(object):
         certFile = tmpdir.join('cert_and_chain.pem')
         certFile.write(''.join(CERT_PEMS))
         with pytest.raises(ValueError):
-            pem.certificateOptionsFromFiles(
+            certificateOptionsFromFiles(
                 str(certFile)
             )
 
@@ -102,7 +104,7 @@ class TestCertificateOptionsFromFiles(object):
         allFile = tmpdir.join('key_cert_and_chain.pem')
         allFile.write(KEY_PEM + ''.join(CERT_PEMS) + KEY_PEM2)
         with pytest.raises(ValueError):
-            pem.certificateOptionsFromFiles(
+            certificateOptionsFromFiles(
                 str(allFile)
             )
 
@@ -110,7 +112,7 @@ class TestCertificateOptionsFromFiles(object):
         keyFile = tmpdir.join('key.pem')
         keyFile.write(KEY_PEM)
         with pytest.raises(ValueError):
-            pem.certificateOptionsFromFiles(
+            certificateOptionsFromFiles(
                 str(keyFile)
             )
 
@@ -122,16 +124,18 @@ class TestCertificateOptionsFromFiles(object):
         keyFile = tmpdir.join('key.pem')
         keyFile.write(KEY_PEM + "".join(CERT_PEMS[1:]))
         with pytest.raises(ValueError) as excinfo:
-            pem.certificateOptionsFromFiles(
+            certificateOptionsFromFiles(
                 str(keyFile)
             )
         assert str(excinfo.value).startswith("No certificate matching ")
 
 
 class TestForwardCompatibleDHE(object):
-    def test_fakeDHParameterSupport(self, monkeypatch, allFile):
+    def test_fakeDHParameterSupport(self, monkeypatch, allFile, recwarn):
         """
         Fake DH parameter support if Twisted doesn't support it.
+
+        Warns about deprecation.
         """
         fakeCtxFactory = object()
         recorder = call_recorder(lambda *a, **kw: fakeCtxFactory)
@@ -139,7 +143,7 @@ class TestForwardCompatibleDHE(object):
         monkeypatch.setattr(pem.twisted, "_DH_PARAMETERS_SUPPORTED", False)
 
         fakeParameters = object()
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(allFile),
             dhParameters=fakeParameters
         )
@@ -147,6 +151,12 @@ class TestForwardCompatibleDHE(object):
         assert isinstance(ctxFactory, pem.twisted._DHParamContextFactory)
         assert ctxFactory.ctxFactory is fakeCtxFactory
         assert "dhParameters" not in recorder.calls[0].kwargs
+
+        w = recwarn.pop(DeprecationWarning)
+        assert (
+            "The backport of DiffieHellmanParameters will be removed."
+            in str(w.message)
+        )
 
     def test_realDHParameterSupport(self, monkeypatch, allFile):
         """
@@ -159,7 +169,7 @@ class TestForwardCompatibleDHE(object):
         monkeypatch.setattr(pem.twisted, "_DH_PARAMETERS_SUPPORTED", True)
 
         fakeParameters = object()
-        ctxFactory = pem.certificateOptionsFromFiles(
+        ctxFactory = certificateOptionsFromFiles(
             str(allFile),
             dhParameters=fakeParameters
         )
@@ -190,3 +200,29 @@ class TestForwardCompatibleDHE(object):
         ctx = ctxFactory.getContext()
         assert fakeContext is ctx
         assert [call(b"foo")] == fakeContext.load_tmp_dh.calls
+
+
+class TestDeprecations(object):
+    def test_certificateOptionsFromFiles(self, monkeypatch, recwarn):
+        """
+        pem.certificateOptionsFromFiles raises a deprecation warning and calls
+        the original method with the same arguments.
+        """
+        cr = call_recorder(lambda *a, **kw: None)
+        monkeypatch.setattr(pem, "certificateOptionsFromFilesOriginal", cr)
+        pem.certificateOptionsFromFiles("foo", bar="baz")
+        assert [call("foo", bar="baz")] == cr.calls
+        w = recwarn.pop(DeprecationWarning)
+        assert "certificateOptionsFromFiles" in str(w.message)
+
+    def test_certificateOptionsFromPEMs(self, monkeypatch, recwarn):
+        """
+        pem.certificateOptionsFromPEMs raises a deprecation warning and calls
+        the original method with the same arguments.
+        """
+        cr = call_recorder(lambda *a, **kw: None)
+        monkeypatch.setattr(pem, "certificateOptionsFromPEMsOriginal", cr)
+        pem.certificateOptionsFromPEMs("foo", bar="baz")
+        assert [call("foo", bar="baz")] == cr.calls
+        w = recwarn.pop(DeprecationWarning)
+        assert "certificateOptionsFromPEMs" in str(w.message)